course controller ok

+++ SistemaAcademico/Controllers/CourseController.cs
@@ -37,51 +37,6 @@ namespace AcademicSystemApi.Controllers
             this._subjectService = subjectService;
         }
 
-        private async Task<bool> PermissionCheckToReadCourse(Course c)
-        {
-            User u = (await _userService.GetByID(this.GetUserID())).Data[0];
-
-            // verify 
-            bool hasPermissionToRead = false;
-            if (u.Owner != null && u.Owner.IsActive)
-            {
-                Owner o = (await _ownerService.GetByID(u.Owner.ID)).Data[0];
-                foreach (OwnerCourse ownerCourse in o.Courses)
-                {
-                    if (c.Owners.Where(owner => owner.CourseID == ownerCourse.OwnerID).Any())
-                    {
-                        hasPermissionToRead = true;
-                    }
-                }
-            }
-            if (u.Student != null)
-            {
-                Student student = (await this._studentService.GetByID(u.Student.ID)).Data[0];
-                foreach (StudentClass studentClass in student.Classes)
-                {
-                    Class Class = (await this._classService.GetByID(studentClass.ClassID)).Data[0];
-                    if (Class.CourseID == c.ID)
-                    {
-                        hasPermissionToRead = true;
-                    }
-                }
-            }
-            if (u.Coordinator != null)
-            {
-                Coordinator coordinator = (await this._coordinatorService.GetByID(u.Coordinator.ID)).Data[0];
-                foreach (CoordinatorClass coordinatorClass in coordinator.Classes)
-                {
-                    Class Class = (await this._classService.GetByID(coordinatorClass.ClassID)).Data[0];
-                    if (Class.Coordinators.Where(c => c.CoordinatorID == coordinator.ID).Any())
-                    {
-                        hasPermissionToRead = true;
-                    }
-                }
-            }
-
-            return hasPermissionToRead;
-        }
-
 
         [HttpGet]
         [Route("{id}")]
@@ -91,7 +46,7 @@ namespace AcademicSystemApi.Controllers
             try
             {
                 DataResponse<Course> response = await _courseService.GetByID(id);
-                if (await this.VerifyPermisionCourse(response.Data[0]))
+                if (await this.PermissionCheckToReadCourse(response.Data[0]))
                 {
                     return this.SendResponse(response);
                 }
@@ -106,17 +61,18 @@ namespace AcademicSystemApi.Controllers
 
         [HttpPost]
         [Authorize]
-        public async Task<object> CreateCourse(Course Course)
+        public async Task<object> CreateCourse(Course course)
         {
             try
             {
-                User user = (await _userService.GetByID(this.GetUserID())).Data[0];
-                if (user.Owner != null && user.Owner.IsActive)
+                if (await this.CheckPermissionToCreateCourse(course))
                 {
-                    Course.Owners.Clear();
-                    int id = (await _courseService.CreateAndReturnId(Course)).Data[0];
-                    Course = (await _courseService.GetByID(id)).Data[0];
-                    return this.SendResponse(await _courseService.AddOwner(Course, user.Owner));
+                    User user = (await _userService.GetByID(this.GetUserID())).Data[0];
+
+                    course.Owners.Clear();
+                    int id = (await _courseService.CreateAndReturnId(course)).Data[0];
+                    course = (await _courseService.GetByID(id)).Data[0];
+                    return this.SendResponse(await _courseService.AddOwner(course, user.Owner));
                 }
                 return Forbid();
             }
@@ -130,22 +86,15 @@ namespace AcademicSystemApi.Controllers
         [HttpPut]
         [Authorize]
         [Route("{id}")]
-        public async Task<object> UpdateCourse(Course Course, int id)
+        public async Task<object> UpdateCourse(Course course, int id)
         {
-            Course.ID = id;
+            course.ID = id;
             try
             {
                 User user = (await _userService.GetByID(this.GetUserID())).Data[0];
-                if (user.Owner != null && user.Owner.IsActive)
+                if (await this.CheckPermissionToUpdateCourse(course))
                 {
-                    Owner owner = (await _ownerService.GetByID(user.Owner.ID)).Data[0];
-                    foreach (OwnerCourse ownerCourse in Course.Owners)
-                    {
-                        if (owner.Courses.Contains(ownerCourse))
-                        {
-                            return this.SendResponse(await _courseService.Update(Course));
-                        }
-                    }
+                    return this.SendResponse(await _courseService.Update(course));
                 }
                 return Forbid();
             }
@@ -163,22 +112,14 @@ namespace AcademicSystemApi.Controllers
         {
             try
             {
-                int ownerID = await this.VerifyIfUserIsOwnerAndReturnOwnerId(_userService);
-                Course course = (await _courseService.GetByID(id)).Data[0];
-                Owner owner = (await _ownerService.GetByID(ownerID)).Data[0];
-                foreach (OwnerCourse ownerCourse in course.Owners)
+                //this.SendResponse(await _courseService.Delete(id));
+                if (await this.CheckPermissionToDeleteCourse(id))
                 {
-                    if (owner.Courses.Contains(ownerCourse))
-                    {
-                        return this.SendResponse(await _courseService.Delete(id));
-                    }
+                    Response response = await _courseService.Delete(id);
+                    return this.SendResponse(response);
                 }
 
-                Response response = await _courseService.Delete(id);
-                return new
-                {
-                    success = response.Success
-                };
+                return Forbid();
             }
             catch (Exception e)
             {
@@ -187,6 +128,67 @@ namespace AcademicSystemApi.Controllers
             }
         }
 
+        [HttpPost]
+        [Route("subject")]
+        [Authorize]
+        public async Task<object> AddSubject(int courseID, int subjectID)
+        {
+            try
+            {
+                if (await this.PermissionCheckToAddSubject(courseID))
+                {
+                    return await _courseService.AddSubject((await _courseService.GetByID(courseID)).Data[0], (await _subjectService.GetByID(subjectID)).Data[0]);
+                }
+
+
+                return Forbid();
+            }
+            catch (Exception e)
+            {
+                Response.StatusCode = StatusCode(500).StatusCode;
+                return null;
+            }
+        }
+
+        [HttpPost]
+        [Route("owner")]
+        [Authorize]
+        public async Task<object> AddOwner(OwnerCourse ownerCourse)
+        {
+            try
+            {
+                if (await this.PermissionCheckToAddOwner(ownerCourse))
+                {
+                    return await _courseService.AddOwner((await _courseService.GetByID(ownerCourse.CourseID)).Data[0], (await _ownerService.GetByID(ownerCourse.OwnerID)).Data[0]);
+                }
+                return Forbid();
+            }
+            catch (Exception e)
+            {
+                Response.StatusCode = StatusCode(500).StatusCode;
+                return null;
+            }
+        }
+
+
+
+        private async Task<bool> CheckPermissionToCreateCourse(Course course)
+        {
+            try
+            {
+                User user = (await _userService.GetByID(this.GetUserID())).Data[0];
+                if (user.Owner != null && user.Owner.IsActive)
+                {
+                    return true;
+                }
+
+                return false;
+            }
+            catch (Exception)
+            {
+                return false;
+            }
+        }
 
         private async Task<bool> VerifyPermisionCourse(Course course)
         {
@@ -238,32 +240,72 @@ namespace AcademicSystemApi.Controllers
             return false;
         }
 
-        [HttpPost]
-        [Route("subject")]
-        [Authorize]
-        public async Task<object> AddSubject(int CourseID, int SubjectID)
+        private async Task<bool> CheckPermissionToDeleteCourse(int id)
+        {
+            try
+            {
+                int ownerID = await this.VerifyIfUserIsOwnerAndReturnOwnerId(_userService);
+                Course course = (await _courseService.GetByID(id)).Data[0];
+                Owner owner = (await _ownerService.GetByID(ownerID)).Data[0];
+                foreach (OwnerCourse ownerCourse in course.Owners)
+                {
+                    if (owner.Courses.Contains(ownerCourse))
+                    {
+                        return true;
+                    }
+                }
+
+                return false;
+            }
+            catch (Exception)
+            {
+                return false;
+            }
+        }
+        private async Task<bool> CheckPermissionToUpdateCourse(Course course)
+        {
+            try
+            {
+                User user = (await _userService.GetByID(this.GetUserID())).Data[0];
+                if (user.Owner != null && user.Owner.IsActive)
+                {
+                    Owner owner = (await _ownerService.GetByID(user.Owner.ID)).Data[0];
+                    foreach (OwnerCourse ownerCourse in course.Owners)
+                    {
+                        if (owner.Courses.Contains(ownerCourse))
+                        {
+                            return true;
+                        }
+                    }
+                }
+                return false; 
+            }
+            catch (Exception)
+            {
+                return false;
+            }
+        }
+
+        private async Task<bool> PermissionCheckToAddSubject(int courseID)
         {
             try
             {
                 User user = (await _userService.GetByID(this.GetUserID())).Data[0];
                 Owner userOwner = (await _ownerService.GetByID(user.Owner.ID)).Data[0];
-                if (userOwner.Courses.Where(c => c.CourseID == CourseID).Any())
+                if (userOwner.Courses.Where(c => c.CourseID == courseID).Any())
                 {
-                    return await _courseService.AddSubject((await _courseService.GetByID(CourseID)).Data[0], (await _subjectService.GetByID(SubjectID)).Data[0]);
+                    return true;
                 }
-                return Forbid();
+
+                return false; 
             }
-            catch (Exception e)
+            catch (Exception)
             {
-                Response.StatusCode = StatusCode(500).StatusCode;
-                return null;
+                return false;
             }
         }
 
-        [HttpPost]
-        [Route("owner")]
-        [Authorize]
-        public async Task<object> AddOwner(OwnerCourse ownerCourse)
+        private async Task<bool> PermissionCheckToAddOwner(OwnerCourse ownerCourse)
         {
             try
             {
@@ -271,15 +313,60 @@ namespace AcademicSystemApi.Controllers
                 Owner userOwner = (await _ownerService.GetByID(user.Owner.ID)).Data[0];
                 if (userOwner.Courses.Where(c => c.CourseID == ownerCourse.CourseID).Any())
                 {
-                    return await _courseService.AddOwner((await _courseService.GetByID(ownerCourse.CourseID)).Data[0] , (await _ownerService.GetByID(ownerCourse.OwnerID)).Data[0]);
+                    return true;
                 }
-                return Forbid();
+
+                return false;
             }
-            catch (Exception e)
+            catch (Exception)
             {
-                Response.StatusCode = StatusCode(500).StatusCode;
-                return null;
+                return false; 
             }
+        }
+
+        private async Task<bool> PermissionCheckToReadCourse(Course c)
+        {
+            User u = (await _userService.GetByID(this.GetUserID())).Data[0];
+
+            // verify 
+            bool hasPermissionToRead = false;
+            if (u.Owner != null && u.Owner.IsActive)
+            {
+                Owner o = (await _ownerService.GetByID(u.Owner.ID)).Data[0];
+                foreach (OwnerCourse ownerCourse in o.Courses)
+                {
+                    if (c.Owners.Where(owner => owner.CourseID == ownerCourse.OwnerID).Any())
+                    {
+                        hasPermissionToRead = true;
+                    }
+                }
+            }
+            if (u.Student != null && u.Student.IsActive)
+            {
+                Student student = (await this._studentService.GetByID(u.Student.ID)).Data[0];
+                foreach (StudentClass studentClass in student.Classes)
+                {
+                    Class Class = (await this._classService.GetByID(studentClass.ClassID)).Data[0];
+                    if (Class.CourseID == c.ID)
+                    {
+                        hasPermissionToRead = true;
+                    }
+                }
+            }
+            if (u.Coordinator != null && u.Coordinator.IsActive)
+            {
+                Coordinator coordinator = (await this._coordinatorService.GetByID(u.Coordinator.ID)).Data[0];
+                foreach (CoordinatorClass coordinatorClass in coordinator.Classes)
+                {
+                    Class Class = (await this._classService.GetByID(coordinatorClass.ClassID)).Data[0];
+                    if (Class.Coordinators.Where(c => c.CoordinatorID == coordinator.ID).Any())
+                    {
+                        hasPermissionToRead = true;
+                    }
+                }
+            }
+
+            return hasPermissionToRead;
         }
     }
 }
